requirements.txt
flask
pymongo
pika
prometheus_client
python-dotenv
pytest
